+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 24/06/2022 - v3.5.1
+
+- **PCB Exposure:**
+   - (Fix) Able to omit X or Y coordinate and use last known value in place
+   - (Fix) Reusing macros in apertures was causing to use the last aperture values for all previous apertures that share the same macro name
+   - (Fix) Implement the inch measurement (%MOIN*%)
+   - (Fix) Crash when redo the operation (Ctrl + Shift + Z)
+- **UI - Issue list:**
+   - (Add) Context menu when right click issues to select an action
+   - (Add) Option to solidify suction cups when right click on the issue
+   - (Improvement) Better confirmation text when click on remove issue(s) with detailed list of actions
+
 ## 19/06/2022 - v3.5.0
 
 - **PCB Exposure:**